refactor(ask_user): extract tool UI into dedicated file
